adding manual extraction of data using contributor ci
the contributor-ci.yaml file includes all radiuss repos, and can be run manually
on the command line, or nightly as the workflow. This data will populate into our new
RADIUSS interface (TBA!). The action currently has a test branch of contributor CI as
we need to test the new ability for the action to customize the output directory format,
changing it to year/month so we do not cache data daily! Finally, I am testing a token
with more limited scopes, likely anticipating it will fail because GraphQL needs higher
permissions, and on that case we will create a more suitable GitHub token.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>